moved all test files to _test_ folder and rename all test files to same as normal components anme
