+++ frontend/src/_test_/Categories.test.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { shallow, configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import Categories from '../../Components/Categories';
+import Categories from '../Components/Categories';
 
 configure({ adapter: new Adapter() });
