+++ frontend/src/_test_/HomeSearch.test.js
@@ -3,8 +3,8 @@ import { shallow, configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 import { Provider } from 'react-redux';
 import { createStore } from 'redux';
-import reducerRoot from '../../reducers';
-import HomeSearch from '../../Components/HomePage/HomeSearch';
+import reducerRoot from '../reducers';
+import HomeSearch from '../Components/HomePage/HomeSearch';
 
 configure({ adapter: new Adapter() });
