+++ frontend/src/_test_/LandingPage.test.js
@@ -3,17 +3,17 @@ import { shallow, configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 import { Provider } from 'react-redux';
 import { createStore } from 'redux';
-import reducerRoot from '../../reducers';
-import HomePage from '../../Components/HomePage';
+import reducerRoot from '../reducers';
+import LandingPage from '../Components/LandingPage';
 
 configure({ adapter: new Adapter() });
 
 const store = createStore(reducerRoot);
 
-describe('HomePage Component', () => {
+describe('LandingPage Component', () => {
     it('Should render without crashing', () => {
         const wrapper = 
-        shallow(<Provider store={store}><HomePage /></Provider>).length
+        shallow(<Provider store={store}><LandingPage /></Provider>).length
         expect(wrapper).toEqual(1)
     })
 })
